# performance displaying graph with large jtl file