fixes-9522

On Macbook pro:
Current implementation:
Large 45 MB test file takes about 7 secs on a file being processed the first time.
Serialized results jtl file of about 27 MB takes  6 secs.

With the fix of using BufferedInputStream and BufferedOutputStream.
Large 45 MB test file takes about 3 secs on a file being processed the first time.
Serialized results jtl file of about 27 MB takes  1.5 secs.